chore: refine miner skeleton
