+++ src/node/miner/mod.rs
@@ -1,3 +1,8 @@
 pub mod payload_builder;
 pub mod util;
 pub mod signer;
+pub mod miner;
+pub mod config;
+
+pub use miner::BscMiner;
+pub use config::{MiningConfig, keystore};
